Add separate files for new modules
