+++ index.html
@@ -7,7 +7,6 @@
   <title>Awesome Books</title>
   <link rel="stylesheet" href="styles.css">
   <link rel="stylesheet" href="https://bootswatch.com/5/flatly/bootstrap.min.css">
-  <script src="app.js" defer></script>
   <script src="https://cdn.jsdelivr.net/npm/luxon@1.27.0/build/global/luxon.min.js"></script>
 </head>
 <body>
@@ -69,5 +68,6 @@
   <footer class="myfoot bg-primary p-2 mt-5 ">
       <p class="m-2">copyright &copy; 2021, All Rights Reserved</p>
   </footer>
+  <script src="app.js" type="module"></script>
 </body>
 </html>
